Refactor: remove unused directory
