+++ src/components/Header/Header.jsx
@@ -4,7 +4,7 @@ import supabase from "../../supabaseClient";
 import { IconMyPage } from "../Icon/Icons/IconMyPage";
 import { Button, ButtonContainer, HeaderBar, HeaderLink, IconContainer, Img, LogoContainer } from "./Header.styled";
 import SearchContainer from "./SearchContainer/SearchContainer";
-import useUserStore from "../../../store/userStore";
+import useUserStore from "../../store/userStore";
 
 const Header = () => {
   const [userName, setUserName] = useState("");
